requirements.txt
fastapi
uvicorn
python-multipart
azure-ai-vision-imageanalysis
azure-core